create handleImageClick function
created handleImageClick function that is used as onClick() in added image rectangles to set file to image clicked and navigate to step 2. As well as conditionally check whether showCheckBoxes are displayed so that it does not implement this onClick() function if checkboxes are displayed so that user is able to check checkboxes to delete image.