Update triangle color and position dynamically based on progress bar color
Updated the MultiStepFormActiveTest component to dynamically set the color of the triangle based on the color of the progress bar. Now, the triangle's borderBottom style property is set inline with the color variable passed from the renderRatingRow function. This ensures consistency in the appearance of the triangle with respect to the progress bar color. As well as point the triangle to the correct  position on progress bars.